refactor: move tests to submodule
This change will significantly lower the amount of dependencies required by the main module.
