+++ go.mod
@@ -4,11 +4,7 @@ go 1.13
 
 require (
 	emperror.dev/errors v0.8.0
-	github.com/imdario/mergo v0.3.7 // indirect
 	github.com/json-iterator/go v1.1.10
 	k8s.io/api v0.19.2
-	k8s.io/apiextensions-apiserver v0.19.2
 	k8s.io/apimachinery v0.19.2
-	k8s.io/client-go v0.19.2
-	k8s.io/klog/v2 v2.3.0
 )
